feat: [#56] transfer user password from v1.0.0 to v2.0.0
We transfer the field `torrust_users.password` to
`torrust_user_authentication.password_hash`.

The hash value is using the PHC string format since v.1.0.0.

In v1.0.0 we were using the hash function "pbkdf2-sha256".
In v2.0.0 we are using "argon2id".

The packages we use to verify password allow using different hash
functions. So we only had to use a different algorithm depending on the
hash id in the PHC string.