fix: [#84] restore behavior. Update torrent stats after upload
When a new torrent is uploaded we have to update the tracker torrent
stats in the `torrust_torrent_tracker_stats` table.

It was from the UI but not for the DB migration script becuase the
frontend makes a request to get the torrent info after the upload and
that endpint updates the torrent tracket stats.

It must update the stats even if that endpoint is not called after
uploading a new torrent.