refactor: run e2e tests with independent isolated servers too.
E2E tests can be executed with a shared app instance with:

```
TORRUST_IDX_BACK_E2E_SHARED=true cargo test
```

or with an isolated instance per test:

```
cargo test
```

In the first case, you have to start the env manbually before running the tests, and manually stop it when tests are finished.
In the second test, the test will launch a new app on a different socket using a different database.

Shared env:

- Requires docker
- It's slower
- It comes with a real traker

Isolated env:

- Doesn't require docker
- It's faster
- You have to mock the TrackerService (not implemented yet)