test: [#130] do not update settings for shared test evns
The test `it_should_allow_admins_to_update_all_the_settings` overwrites
the config file `config.toml`. It uses the same values but:

1. That does not assert that the file was actually updated.
2. It can lead to conflicts with other tests since all of them use the
   same shared env.

For isolated environments, it should not be a problem becuase we inject
the configuration into the app directly wuthout getting it from the
environment via confif file or env var.