fix: failing test
The test for listing torrents was failing becuase the uploaded torrents
was not in the result. Since we added a pagination limit in this PR:

https://github.com/torrust/torrust-index-backend/pull/142

Now it can happen that the torrent you've just uploaded is not in the
response, because tests run in paralell.

When we will able to run this test mocking the `TrackerService` we
couuld run it with an isolated environment, so the expected result if
always only one torrent. Then we could check that the torrent is on the
list again.