feat!: change random torrent generator endpoint
The new enpopint is:

`http://0.0.0.0:3001/v1/torrent/meta-info/random/:uuid`

The segments have changed to differenciate the indexed torrent from the torrent file (meta-indo).

An indexed torrent is a torrent file (meta-info file) with some extra
classification metadata: title, description, category and tags.

There is also a new PATH param `:uuid` which is an UUID to identify the
generated torrent file. The UUID is used for:

- The torrent file name
- The sample contents for a text file ffrom which we generate the
  torrent file.