fix: [#242] wrong infohash when info dict contains source field
When you define a "source" field value in the "info" dictionary inside
the torrent file the field changes the infohash value. We did not
save that field in the database and in the in-memory struct
`TorrentInfo` so the calculated infohash was wrong becuase this field
belongs to the `info` key.