feat: [#2560] new function to calculate the original torrent infohash
We can change the final torrent infohash becuase we do not extract/parse
non standard fields in the info dictionary. We want to keep a copy of
the original uploaded torrent's infohash. Si we need to calcuylate the
infohash with all the fields in the info dictionary.

We will sotre it in the database and warn users in the UI  when the infohash
changes.