feat!: [#288] do not allow empty category names
It was allowed to use an empty string like "" or " " for a category name.

From now on, it's not allowed.

If there were some empty names in the database, they were not renamed. Admins must optionally do that. Names were anyway UNIQUE.

A migration to rename empty names was not added because there can be more than one category, for example:

- ""
- " "
- "  "
- Etcetera

We could have generated names like "no category 1", "no category 2", but it's not likely that admins have created empty categories.