feat: [#483] allow upload torrent with application/octet-stream
HTTP header Content-Type.

Some users are having problems uploading torrents becuase the client
(browser or other clients) use the HTTP header Contetnt-Type:

`application/octet-stream`

instead of:

`application/x-bittorrent`

It seems that the reason is they don't have any application associated
to that extension. So it uses the generic binary file mime type.

The MIME type can be inferred from the file extension. If the system or application uploading the file has a specific association for .torrent files, it might set the MIME type to application/x-bittorrent. In the absence of such association, it might fall back to the generic application/octet-stream.