fix: [#581] broken tests using relative paths for fixtures
after introducing figment::Jail for testing. It seems Jail changes the
current dir and that make other tests fails if they use relative path to
load files (in this case fixtures).

A solution like useing the env var CARGO_MANIFEST_DIR does not work with
`cargo next` when building the container. It only worked when you run
the tests with cargo test.