refactor: [#599] use URL type for connect_url in database config
We were using a cusomt type but sqlx defines it as a URL. WE can
validate this type in the configration and let the app domain validate
if the connection URL is a valid one. For example, if the app supports
the database driver. Otherwise we have to change this first level
valitation anytime we support a new DB.