test: [#628] add failing test for bug. Duplicate slash in announce URL
When the tracker is running in private mode the announce URL in the
torrent file includes a double slash '/'. For example:

http://tracker:7070//KEY

This commit adds a failing test.

It needed scaffolding to run E2E tests with a tracker running in
"private" mode. We have been only running E2E tests for public trackers so
far. The new tests is only execute with Sqlite3. It would be easy to run
it also for MySQL. It was not cinluded in order to avoid making tests to
slow. It requires to start the docker E2E shared environment twice (for
public and private tracker configurations).