refactor: move ConfigurationPublic
ConfigurationPublic is not a configuration type. It does not belong to
the configuration. We don't need to version this type when the
configuration changes.

It's a type containing a subset of the data contained in the
configuration that we want to expose via the API. It happens that those
are the fields we want to expose via the API becuase they are the fields
we are using in the webapp (Index GUI) but it's not part of the
configuration. It's a concrete view of the configration used for other
purposes rather than initialize the Index app.

In the future, it could be even moved to the API as a API resource.

Changing this struct changes the API contract. The contract with the API
consumers, not the contract with the Index administrators, the people
responsible for setting up the Index and it's configuration.

That the reason why it was moved from the config mod to the config
service.

It's not a problem now, but we should cerate an API resource for this
type becuase it should be versioned in the API. WE are using versioning
in the API but the type was excluded, meaning it cannot be versioned.