feat: [#730] add CMS content to configration
We need to pass legal texts to the frontent but we don't have a proper
CMS. For now, we have just added the texts in the configuration to make
it possible for users to overrride them. These are the new values in the
config:

```toml
[website.demo]
warning = "⚠️ Please be aware: This demo resets all data weekly. Torrents not complying with our Usage Policies will be removed immediately without notice. We encourage the responsible use of this software in compliance with all legal requirements."

[website.terms]

[website.terms.upload]
content_upload_agreement = """I confirm that the content I am uploading is authorized, and I have read and agree to the terms."""

[website.terms.page]
content = """

Our software is designed to support the distribution of legal, authorized content only. Users may only upload or share files that fall under the following categories:

- **Open-Source Licenses:** Content licensed under recognized open-source licenses, allowing for free distribution and modification.
- **Creative Commons Licenses:** Content released under Creative Commons licenses that permit sharing and distribution.
- **Public Domain:** Content that is free of copyright restrictions and available for public use.

**Prohibited Content:** Any content that infringes copyright, is subject to copyright protection, or is illegal under applicable laws is strictly prohibited. This includes but is not limited to copyrighted movies, music, software, books, and any other media.

**Enforcement:** We reserve the right to remove any content that does not comply with these policies without notice. We may also take additional steps, including reporting violations to the relevant authorities, if necessary.

"""
title = "Usage Policies and Content Restrictions"
```