fix: chedck error
Mutable static variables are discouraged. I guess we should move them to
the proxy service. We wanted to avoid generating the error images every
time we need them becuase it's costly. I would created them when the
proxy service is instanciated or maybe we can create a new service with
static content for the proxy.